requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
extruct==0.16.0
html-text==0.5.2
html5lib==1.1
idna==3.4
install==1.3.5
isodate==0.6.1
jstyleson==0.0.2
lxml==4.9.3
mf2py==1.1.3
psycopg2-binary==2.9.7
pyparsing==3.1.1
pyRdfa3==3.5.3
rdflib==7.0.0
recipe-scrapers==14.41.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
urllib3==2.0.4
w3lib==2.1.2
webencodings==0.5.1
